refactor: Health check endpoints now in api folder, logger config at project root. Lifespan in db folders.
